# doc/api.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: WASAPhoto
  description: |-
    Each user will be presented with a stream of photos (images) in reverse chronological order, with information about when each photo was uploaded (date and time) and how many likes and comments it has. The stream is composed by photos from “following” (other users that the user follows). Users can place (and later remove) a “like” to photos from other users. Also, users can add comments to any image (even those uploaded by themself). Only authors can remove their comments. 
    
    Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about Alice. Alice can decide to remove the ban at any moment. 
    
    Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse chronological order), how many photos have been uploaded, and the user’s followers and following. Users can change their usernames, upload photos, remove photos, and follow/unfollow other users. Removal of an image will also remove likes and comments. 
    
    A user can search other user profiles via username. 
    
    A user can log in just by specifying the username.
  
  contact:
    name: Michelangelo Crea
    email: crea.1993024@studenti.uniroma1.it
    url: https://github.com/mikusander
...
paths:
...
  /users/{username}/follow/{followid}:
    parameters: 
      - name: username
        in: path
        required: true
        description: user username
        schema: {$ref: "#/components/schemas/User/properties/username"}
      - name: followid
        in: path
        required: true
        description: follow user username
        schema: {$ref: "#/components/schemas/User/properties/username"}

    put:
      security:
        - bearerAuth: []
      tags: 
        - follow
      summary: Follow user
      description: |
        this endpoint is used to follow another user.
      operationId: followUser
      responses:
        "200":
          description: follow user successfully.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/follow" }
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }
    
    delete:
      security:
        - bearerAuth: []
      tags: 
        - follow
      summary: unfollow user
      description: |
        this endppoint is used to unfollow a user
      operationId: unfollowUser
      responses:
        "204":
          description: unfollow user successfully
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/ban/{banid}:
    parameters: 
      - name: username
        in: path
        required: true
        description: User username
        schema: {$ref: "#/components/schemas/User/properties/username"}
      - name: banid
        in: path
        required: true
        description: user username to ban
        schema: {$ref: "#/components/schemas/User/properties/username"}

    put:
      security:
        - bearerAuth: []
      tags: 
        - ban
      summary: Ban user
      description: |
        This endpoint is used to ban a user.
      operationId: banUser
      responses:
        "200":
          description: ban user successfully.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ban" }
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      security:
        - bearerAuth: []
      tags: 
        - ban
      summary: Unban user
      description: |
        This endpoint is used to unban a user.
      operationId: unbanUser
      responses:
        "204":
          description: user unban successfully
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/photo/{photoid}/like/{likeid}:
    parameters: 
      - name: username
        in: path
        required: true
        description: user username
        schema: {$ref: "#/components/schemas/User/properties/username"}
      - name: photoid
        in: path
        required: true
        description: id photo to be able to been a like
        schema: {$ref: '#/components/schemas/post_photo/properties/id'}
      - name: likeid
        in: path
        required: true
        description: user username that receives like
        schema: {$ref: "#/components/schemas/User/properties/username"}

    put:
      security:
        - bearerAuth: []
      tags: 
        - like
      summary: Like photo
      description: |
        This endpoint is used to put like.
      operationId: likePhoto
      responses:
        "200":
          description: like photo successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/like" }
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }
   
    delete:
      security:
        - bearerAuth: []
      tags: 
        - like
      summary: Unlike photo
      description: |
        This endpoint is used to unlike a photo
      operationId: unlikePhoto
      responses:
        "204":
          description: Like removed
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/photo/{photoid}/comment:
    parameters: 
      - name: username
        in: path
        required: true
        description: user username
        schema: {$ref: "#/components/schemas/User/properties/username"}
      - name: photoid
        in: path
        required: true
        description: photo id
        schema: {$ref: '#/components/schemas/post_photo/properties/id'}
    post:
      security:
        - bearerAuth: []
      tags: 
        - comment
      summary: comment photo
      description: |
        This endpoint is used to put a comment to a photo
      operationId: commentPhoto
      requestBody: 
        content:
          application/json: 
            schema:
              { $ref: "#/components/schemas/comment" }
      responses:
        "201":
          description: photo comment successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/comment" }
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/photo/{photoid}/comment/{commentid}:
    parameters: 
      - name: username
        in: path
        required: true
        description: User username
        schema: {$ref: "#/components/schemas/User/properties/username"}
      - name: photoid
        in: path
        required: true
        description: photo id
        schema: {$ref: '#/components/schemas/post_photo/properties/id'}
      - name: commentid
        in: path
        required: true
        description: comment id
        schema: {$ref: "#/components/schemas/comment/properties/id"}
          
    delete:
      security:
        - bearerAuth: []
      tags: 
        - comment
      summary: uncomment photo
      description: |
        This endpoint is used to remove a comment.
      operationId: uncommentPhoto
      responses:
        "204":
          description: the comment is removed
        "400": { $ref: '#/components/responses/BadRequest' }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses: 
...
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
  schemas:
...
    like:
      title: like
      description: this object represent a like of the user
      type: object
      properties:
        personal_user_id:
          description: unique personal ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        owner_photo:
          description: unique personal ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        photo_id:
          description: unique photo ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
    comment:
      title: comment
      description: this object represent a comment of the user
      type: object
      properties:
        id:
          description: unique follow identifier
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          readOnly: true
          example: cj3kl
        text:
          description: text of the comment
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        personal_user_id:
          description: unique personal ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        owner_photo:
          description: unique personal ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        photo_id:
          description: unique photo ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        date:
          description: date of the comment
          type: string
          format: date-time
          example: "2023-11-25T00:00:00Z"
          minLength: 0
          maxLength: 20
...
    ban:
      title: ban
      description: "this object represent a ban user"
      type: object
      properties:
        personal_user_id:
          description: unique personal ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        ban_user_id:
          description: unique follow ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
    follow:
      title: follow
      description: "this object represent a follow user"
      type: object
      properties:
        personal_user_id:
          description: unique personal ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
        follow_user_id:
          description: unique follow ID
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: cj3kl
...